Multiple camera. Warning: cam editor does not work!
git-svn-id: https://svn.code.sf.net/p/deling/code/trunk@66 d07ef2b7-bbf6-4ce4-b825-8d9644e34689

+++ widgets/WalkmeshWidget.h
@@ -51,7 +51,7 @@ private:
 	QWidget *buildGatewaysPage();
 	QWidget *buildMovieCameraPage();
 	void editCaVector(int id, const Vertex_s &values);
-	void editCaPos(int id, int value);
+	void editCaPos(int id, double value);
 	void editExitPoint(int id, const Vertex_s &values);
 	void editDoorPoint(int id, const Vertex_s &values);
